// Application/Frontend/frontend-app/src/generated/auth-admin-client/model/health-details-response.ts
/**
 * Auth Service, health routes
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { HealthDetailsResponseComponents } from './health-details-response-components';
import { HealthStatus } from './health-status';


export interface HealthDetailsResponse { 
    status: HealthStatus;
    checkedAt: string;
    /**
     * How long this process has been running.
     */
    uptimeSeconds: number;
    components: HealthDetailsResponseComponents;
}
export namespace HealthDetailsResponse {
}